+++ tests/features/test_reuse.py
@@ -13,6 +13,7 @@ class Engine:
         self.config = config
 
 
+@dg.node(reuse=False)
 class Database:
     def __init__(self, engine: Engine):
         self.engine = engine
@@ -27,6 +28,7 @@ def test_reuse():
     service1 = dg.resolve(Service)
     service2 = dg.resolve(Service)
     assert id(service2) != id(service1)
+    assert id(service2.database) != id(service1.database)
 
     @dg.node(reuse=True)
     def service_factory() -> Service:
@@ -37,3 +39,4 @@ def test_reuse():
     service3 = dg.resolve(Service)
     service4 = dg.resolve(Service)
     assert id(service3) == id(service4)
+    assert id(service2.database) != id(service1.database)